IMPORTANT: Blocks is now licensed under the MIT license
What does mean for you, as a user? Nothing much. For a developer? All
code from this point on is licensed under the MIT, and if you wish to
use the code, you have to abide by the rules of the MIT. All previous
versions of Blocks, however, are licensed under the GPL V3, so if you
use that code, you must abide by the rules of that license. :)